API-1491: Fix comments
