+++ src/Akeneo/Connectivity/Connection/front/tests/src/webhook/components/EventLogBadge.test.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import '@testing-library/jest-dom/extend-expect';
 import {EventLogBadge} from '@src/webhook/components/EventLogBadge';
-import {render, screen} from 'akeneo-design-system/lib/storybook/test-util';
+import {screen} from '@testing-library/react';
 import {EventSubscriptionLogLevel} from '@src/webhook/model/EventSubscriptionLogLevel';
+import {renderWithProviders} from '../../../test-utils';
 
 test.each([
     [EventSubscriptionLogLevel.INFO, 'INFO', 'rgb(61, 107, 69)'],
@@ -10,7 +11,7 @@ test.each([
     [EventSubscriptionLogLevel.WARNING, 'WARNING', 'rgb(149, 108, 37)'],
     [EventSubscriptionLogLevel.ERROR, 'ERROR', 'rgb(127, 57, 47)'],
 ])('It render the badge with the level %i', (level, label, color) => {
-    render(<EventLogBadge level={level}>{label}</EventLogBadge>);
+    renderWithProviders(<EventLogBadge level={level}>{label}</EventLogBadge>);
 
     const badge = screen.getByText(label);
